requirements.txt
# Production-ready dependencies for Azure App Service
PyYAML>=6.0
Flask>=2.3.0
Werkzeug>=2.3.0
click>=8.0.0
jinja2>=3.1.0
python-dateutil>=2.8.0

# Production WSGI server
gunicorn>=21.2.0
gevent>=23.0.0

# AI Integration
openai>=1.3.0
pydantic>=2.0.0

# Multi-user authentication
authlib>=1.2.0
python-dotenv>=1.0.0